requirements.txt
camelot-py==0.11.0
ghostscript
pdfplumber==0.9.0
loguru==0.7.0
langchain==0.0.247
transformers==4.30.2
sentencepiece==0.1.99
opencv-python
parse
fastbm25
requests
pandas
numpy
matplotlib
text2vec
torch
huggingface_hub
